% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListFromDir.R
\name{get_list_from_directory}
\alias{get_list_from_directory}
\title{Get a file and directory list for ShinyFileTree}
\usage{
get_list_from_directory(my_dir, pattern = NULL,
  hide_empty_dirs = FALSE, show_hidden_files = FALSE,
  hide_files = FALSE, state = NULL, simplify = FALSE,
  max_depth = 10)
}
\description{
Get a file and directory list for ShinyFileTree
}
