% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFileTree.R
\name{shinyFileTree}
\alias{shinyFileTree}
\title{<Add Title>}
\usage{
shinyFileTree(data, plugins = NULL, core_opts = list(check_callback =
  TRUE), is_directory = FALSE, default_plugin_opts = list(checkbox =
  list(visible = TRUE, three_state = TRUE, whole_node = TRUE, cascade =
  TRUE), contextmenu = list(select_node = TRUE, show_at_node = TRUE), types
  = list(file = list(icon = "glyphicon glyphicon-flash"))), ...,
  width = NULL, height = NULL)
}
\description{
<Add Description>
}
\details{
checkbox: https://www.jstree.com/api/#/?q=$.jstree.defaults.checkbox&f=$.jstree.defaults.checkbox.visible
}
