% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFileTree.R
\name{shinyFileTreeOpts}
\alias{shinyFileTreeOpts}
\title{Function to set options for shinyFileTree}
\usage{
shinyFileTreeOpts(animation = 200, multiple = FALSE,
  force_text = FALSE, dblclick_toggle = TRUE, check_callback = TRUE,
  themes.name = FALSE, themes.url = FALSE, themes.dir = FALSE,
  themes.dots = TRUE, themes.icons = TRUE, themes.stripes = TRUE,
  themes.variant = FALSE, themes.reposonsive = FALSE, ...)
}
\arguments{
\item{animation}{the open / close animation duration in milliseconds - set this to false to disable the animation (default is 200)}

\item{multiple}{a boolean indicating if multiple nodes can be selected (default FALSE).}

\item{force_text}{Force node text to plain text (and escape HTML) (default FALSE).}

\item{dblclick_toggle}{Should the node be toggled if the text is double clicked (default TRUE).}

\item{themes.name}{the name of the theme to use (if left as FALSE the default theme is used)}

\item{themes.url}{the URL of the theme's CSS file, leave this as false if you have manually included the theme CSS (recommended). You can set this to true too which will try to autoload the theme.}

\item{themes.dir}{the location of all jstree themes - only used if url is set to TRUE}

\item{themes.dots}{a boolean indicating if connecting dots are shown.}

\item{themes.icons}{a boolean indicating if node icons are shown.}

\item{themes.stripes}{a boolean indicating if the tree background is striped.}

\item{themes.variant}{a string (or boolean false) specifying the theme variant to use (if the theme supports variants), e.g. 'large'.}

\item{themes.reposonsive}{a boolean specifying if a reponsive version of the theme should kick in on smaller screens (if the theme supports it). Defaults to FALSE.}

\item{...}{Additional jstree options}
}
\description{
Use as argument to opts parameter of shinyFileTree
}
