% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFileTree.R
\name{shinyFileTree}
\alias{shinyFileTree}
\title{Create a Shiny File Tree}
\usage{
shinyFileTree(data, plugins = NULL, opts = shinyFileTreeOpts(),
  default_plugin_opts = list(checkbox = list(visible = TRUE, three_state
  = TRUE, whole_node = TRUE, cascade = TRUE), contextmenu =
  list(select_node = TRUE, show_at_node = TRUE), search =
  list(case_sensitive = FALSE, show_only_matches = TRUE), types = list(file
  = list(icon = "glyphicon glyphicon-flash"))), ...,
  is_directory = FALSE, directory_depth = NULL,
  show_dir_info = FALSE, width = NULL, height = NULL)
}
\arguments{
\item{data}{}

\item{plugins}{}

\item{opts}{}

\item{default_plugin_opts}{}

\item{...}{}

\item{is_directory}{}

\item{width}{}

\item{height}{}
}
\description{
This creates a widget with a shiny file tree. The widget is based on jstree,
and the R implementation has the same implementation as jstree. For more 
details, please refer to https://www.jstree.com/docs/config/ and
https://www.jstree.com/api/#/?q=$.jstree.defaults
}
\examples{
shinyFileTree(system.file(package="shinyFileTree"), 
             is_directory = TRUE,
             plugins = c("checkbox"),
             multiple = TRUE,
             opts = shinyFileTreeOpts(icons = TRUE)
)
}
